+++ app/src/main/java/au/com/shiftyjelly/pocketcasts/ui/MainActivity.kt
@@ -33,6 +33,7 @@ import au.com.shiftyjelly.pocketcasts.localization.helper.LocaliseHelper
 import au.com.shiftyjelly.pocketcasts.models.entity.Episode
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.to.SignInState
+import au.com.shiftyjelly.pocketcasts.models.type.EpisodeViewSource
 import au.com.shiftyjelly.pocketcasts.navigation.BottomNavigator
 import au.com.shiftyjelly.pocketcasts.navigation.FragmentInfo
 import au.com.shiftyjelly.pocketcasts.navigation.NavigatorAction
@@ -58,6 +59,7 @@ import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager.PlaybackSource
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackState
 import au.com.shiftyjelly.pocketcasts.repositories.playback.UpNextQueue
+import au.com.shiftyjelly.pocketcasts.repositories.playback.UpNextSource
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PlaylistManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
@@ -125,6 +127,7 @@ class MainActivity :
 
     companion object {
         private const val INITIAL_KEY = "initial"
+        private const val SOURCE_KEY = "source"
         init {
             AppCompatDelegate.setCompatVectorFromResourcesEnabled(true)
         }
@@ -431,11 +434,16 @@ class MainActivity :
     }
 
     override fun onUpNextClicked() {
-        showBottomSheet(UpNextFragment())
+        showUpNextFragment(UpNextSource.MINI_PLAYER)
     }
 
     override fun onMiniPlayerLongClick() {
-        MiniPlayerDialog(playbackManager, podcastManager, episodeManager, supportFragmentManager).show(this)
+        MiniPlayerDialog(playbackManager, podcastManager, episodeManager, supportFragmentManager, analyticsTracker).show(this)
+    }
+
+    private fun showUpNextFragment(source: UpNextSource) {
+        analyticsTracker.track(AnalyticsEvent.UP_NEXT_SHOWN, mapOf(SOURCE_KEY to source.analyticsValue))
+        showBottomSheet(UpNextFragment.newInstance(source = source))
     }
 
     @OptIn(DelicateCoroutinesApi::class)
@@ -522,6 +530,7 @@ class MainActivity :
 
                 override fun onStateChanged(bottomSheet: View, newState: Int) {
                     if (newState == BottomSheetBehavior.STATE_COLLAPSED) {
+                        analyticsTracker.track(AnalyticsEvent.UP_NEXT_DISMISSED)
                         supportFragmentManager.findFragmentByTag(bottomSheetTag)?.let {
                             removeBottomSheetFragment(it)
                         }
@@ -580,7 +589,7 @@ class MainActivity :
         if (intent.getStringExtra(INTENT_EXTRA_PAGE) == "upnext") {
             intent.putExtra(INTENT_EXTRA_PAGE, null as String?)
             binding.playerBottomSheet.openPlayer()
-            onUpNextClicked()
+            showUpNextFragment(UpNextSource.UP_NEXT_SHORTCUT)
         }
     }
 
@@ -646,6 +655,8 @@ class MainActivity :
         frameBottomSheetBehavior.state = BottomSheetBehavior.STATE_COLLAPSED
         binding.frameBottomSheet.importantForAccessibility = View.IMPORTANT_FOR_ACCESSIBILITY_NO
     }
+
+    override fun isUpNextShowing() = bottomSheetTag == UpNextFragment::class.java.name
 
     private fun removeBottomSheetFragment(fragment: Fragment) {
         val tag = fragment::class.java.name
@@ -811,7 +822,12 @@ class MainActivity :
                 // intents were being reused for notifications so we had to use the extra to pass action
                 val episodeUuid =
                     intent.extras?.getString(Settings.INTENT_OPEN_APP_EPISODE_UUID, null)
-                openEpisodeDialog(episodeUuid, null, forceDark = false)
+                openEpisodeDialog(
+                    episodeUuid = episodeUuid,
+                    source = EpisodeViewSource.NOTIFICATION,
+                    podcastUuid = null,
+                    forceDark = false
+                )
             } else if (action == Intent.ACTION_VIEW) {
                 val extraPage = intent.extras?.getString(INTENT_EXTRA_PAGE, null)
                 if (extraPage != null) {
@@ -933,7 +949,12 @@ class MainActivity :
         addFragment(PodcastFragment.newInstance(podcastUuid = uuid))
     }
 
-    override fun openEpisodeDialog(episodeUuid: String?, podcastUuid: String?, forceDark: Boolean) {
+    override fun openEpisodeDialog(
+        episodeUuid: String?,
+        source: EpisodeViewSource,
+        podcastUuid: String?,
+        forceDark: Boolean
+    ) {
         episodeUuid ?: return
 
         launch(Dispatchers.Main.immediate) {
@@ -943,13 +964,15 @@ class MainActivity :
                 val podcastUuidFound = podcastUuid ?: return@launch
                 // Assume it's an episode we don't know about
                 EpisodeFragment.newInstance(
-                    episodeUuid,
+                    episodeUuid = episodeUuid,
+                    source = source,
                     podcastUuid = podcastUuidFound,
                     forceDark = forceDark
                 )
             } else if (playable is Episode) {
                 EpisodeFragment.newInstance(
-                    episodeUuid,
+                    episodeUuid = episodeUuid,
+                    source = source,
                     podcastUuid = podcastUuid,
                     forceDark = forceDark
                 )
@@ -1021,7 +1044,12 @@ class MainActivity :
 
                     val episode = result.episode
                     if (episode != null) {
-                        openEpisodeDialog(episode.uuid, podcastUuid, forceDark = false)
+                        openEpisodeDialog(
+                            episodeUuid = episode.uuid,
+                            source = EpisodeViewSource.SHARE,
+                            podcastUuid = podcastUuid,
+                            forceDark = false
+                        )
                     } else {
                         openPodcastPage(podcastUuid)
                     }