+++ modules/features/filters/src/main/java/au/com/shiftyjelly/pocketcasts/filters/FilterEpisodeListFragment.kt
@@ -21,6 +21,7 @@ import au.com.shiftyjelly.pocketcasts.localization.extensions.getStringPluralPod
 import au.com.shiftyjelly.pocketcasts.models.entity.Episode
 import au.com.shiftyjelly.pocketcasts.models.entity.Playable
 import au.com.shiftyjelly.pocketcasts.models.entity.Playlist
+import au.com.shiftyjelly.pocketcasts.models.type.EpisodeViewSource
 import au.com.shiftyjelly.pocketcasts.podcasts.view.components.PlayButton
 import au.com.shiftyjelly.pocketcasts.podcasts.view.episode.EpisodeFragment
 import au.com.shiftyjelly.pocketcasts.podcasts.view.podcast.EpisodeListAdapter
@@ -156,7 +157,7 @@ class FilterEpisodeListFragment : BaseFragment() {
 
     private fun onRowClick(episode: Playable) {
         if (episode is Episode) {
-            val fragment = EpisodeFragment.newInstance(episode)
+            val fragment = EpisodeFragment.newInstance(episode = episode, source = EpisodeViewSource.FILTERS)
             fragment.show(parentFragmentManager, "episode_card")
         }
     }
@@ -478,23 +479,23 @@ class FilterEpisodeListFragment : BaseFragment() {
                 .setTitle(getString(LR.string.sort_by))
                 .addCheckedOption(
                     titleId = LR.string.episode_sort_newest_to_oldest,
-                    click = { viewModel.changeSort(Playlist.PLAYLIST_SORT_NEWEST_TO_OLDEST) },
-                    checked = (it.sortId == Playlist.PLAYLIST_SORT_NEWEST_TO_OLDEST)
+                    click = { viewModel.changeSort(Playlist.SortOrder.NEWEST_TO_OLDEST) },
+                    checked = (it.sortOrder() == Playlist.SortOrder.NEWEST_TO_OLDEST)
                 )
                 .addCheckedOption(
                     titleId = LR.string.episode_sort_oldest_to_newest,
-                    click = { viewModel.changeSort(Playlist.PLAYLIST_SORT_OLDEST_TO_NEWEST) },
-                    checked = (it.sortId == Playlist.PLAYLIST_SORT_OLDEST_TO_NEWEST)
+                    click = { viewModel.changeSort(Playlist.SortOrder.OLDEST_TO_NEWEST) },
+                    checked = (it.sortOrder() == Playlist.SortOrder.OLDEST_TO_NEWEST)
                 )
                 .addCheckedOption(
                     titleId = LR.string.episode_sort_short_to_long,
-                    click = { viewModel.changeSort(Playlist.PLAYLIST_SORT_SHORTEST_TO_LONGEST) },
-                    checked = (it.sortId == Playlist.PLAYLIST_SORT_SHORTEST_TO_LONGEST)
+                    click = { viewModel.changeSort(Playlist.SortOrder.SHORTEST_TO_LONGEST) },
+                    checked = (it.sortOrder() == Playlist.SortOrder.SHORTEST_TO_LONGEST)
                 )
                 .addCheckedOption(
                     titleId = LR.string.episode_sort_long_to_short,
-                    click = { viewModel.changeSort(Playlist.PLAYLIST_SORT_LONGEST_TO_SHORTEST) },
-                    checked = (it.sortId == Playlist.PLAYLIST_SORT_LONGEST_TO_SHORTEST)
+                    click = { viewModel.changeSort(Playlist.SortOrder.LONGEST_TO_SHORTEST) },
+                    checked = (it.sortOrder() == Playlist.SortOrder.LONGEST_TO_SHORTEST)
                 )
             dialog.show(parentFragmentManager, "sort_options")
         }