+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/NotesFragment.kt
@@ -14,6 +14,8 @@ import android.widget.Toast
 import androidx.core.os.bundleOf
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.player.databinding.FragmentNotesBinding
 import au.com.shiftyjelly.pocketcasts.player.viewmodel.NotesViewModel
 import au.com.shiftyjelly.pocketcasts.player.viewmodel.PlayerViewModel
@@ -37,6 +39,7 @@ class NotesFragment : BaseFragment() {
 
     @Inject lateinit var settings: Settings
     @Inject lateinit var playbackManager: PlaybackManager
+    @Inject lateinit var analyticsTracker: AnalyticsTrackerWrapper
 
     private val playerViewModel: PlayerViewModel by activityViewModels()
     private val viewModel: NotesViewModel by viewModels()
@@ -45,6 +48,8 @@ class NotesFragment : BaseFragment() {
 
     companion object {
         internal const val ARG_EPISODE_UUID = "episode_uuid"
+        private const val EPISODE_UUID_KEY = "episode_uuid"
+        private const val UNKNOWN = "unknown"
 
         fun newInstance(episodeUuid: String): NotesFragment {
             return NotesFragment().apply {
@@ -120,6 +125,7 @@ class NotesFragment : BaseFragment() {
                             jumpToTime(time)
                             return true
                         }
+                        analyticsTracker.track(AnalyticsEvent.PLAYER_SHOW_NOTES_LINK_TAPPED, mapOf(EPISODE_UUID_KEY to (viewModel.episode.value?.uuid ?: UNKNOWN)))
                         return IntentUtil.webViewShouldOverrideUrl(url, view.context)
                     }