+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/PlaybackManager.kt
@@ -107,7 +107,12 @@ open class PlaybackManager @Inject constructor(
         private const val MAX_TIME_WITHOUT_FOCUS_FOR_RESUME_MINUTES = 30
         private const val MAX_TIME_WITHOUT_FOCUS_FOR_RESUME = (MAX_TIME_WITHOUT_FOCUS_FOR_RESUME_MINUTES * 60 * 1000).toLong()
         private const val PAUSE_TIMER_DELAY = ((MAX_TIME_WITHOUT_FOCUS_FOR_RESUME_MINUTES + 1) * 60 * 1000).toLong()
-        private const val KEY_SOURCE = "source"
+        private const val SOURCE_KEY = "source"
+        private const val SEEK_TO_PERCENT_KEY = "seek_to_percent"
+        private const val SEEK_FROM_PERCENT_KEY = "seek_from_percent"
+        const val SPEED_KEY = "speed"
+        const val AMOUNT_KEY = "amount"
+        const val ENABLED_KEY = "enabled"
     }
 
     override val coroutineContext: CoroutineContext
@@ -1846,8 +1851,41 @@ open class PlaybackManager @Inject constructor(
             Timber.w("Found unknown playback source.")
         }
         if (!playbackSource.skipTracking()) {
-            analyticsTracker.track(event, mapOf(KEY_SOURCE to playbackSource.analyticsValue))
+            analyticsTracker.track(event, mapOf(SOURCE_KEY to playbackSource.analyticsValue))
         }
+    }
+
+    fun trackPlaybackSeek(
+        positionMs: Int,
+        playbackSource: PlaybackSource
+    ) {
+        val episode = getCurrentEpisode()
+        episode?.let {
+            val fromPositionMs = episode.playedUpToMs.toDouble()
+            val durationMs = episode.duration * 1000
+            val seekFromPercent = ((fromPositionMs / durationMs) * 100).toInt()
+            val seekToPercent = ((positionMs / durationMs) * 100).toInt()
+
+            analyticsTracker.track(
+                AnalyticsEvent.PLAYBACK_SEEK,
+                mapOf(
+                    SOURCE_KEY to playbackSource.analyticsValue,
+                    SEEK_FROM_PERCENT_KEY to seekFromPercent,
+                    SEEK_TO_PERCENT_KEY to seekToPercent
+                )
+            )
+        }
+    }
+
+    fun trackPlaybackEffectsEvent(
+        event: AnalyticsEvent,
+        props: Map<String, Any> = emptyMap(),
+        playbackSource: PlaybackSource
+    ) {
+        val properties = HashMap<String, Any>()
+        properties[SOURCE_KEY] = playbackSource.analyticsValue
+        properties.putAll(props)
+        analyticsTracker.track(event, properties)
     }
 
     enum class PlaybackSource(val analyticsValue: String) {
@@ -1871,6 +1909,8 @@ open class PlaybackManager @Inject constructor(
         CHROMECAST("chromecast"),
         AUTO_PLAY("auto_play"),
         AUTO_PAUSE("auto_pause"),
+        PLAYER_PLAYBACK_EFFECTS("player_playback_effects"),
+        PODCAST_SETTINGS("podcast_settings"),
         UNKNOWN("unknown");
 
         fun skipTracking() = this in listOf(AUTO_PLAY, AUTO_PAUSE)