+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/PlaylistManagerImpl.kt
@@ -2,6 +2,8 @@ package au.com.shiftyjelly.pocketcasts.repositories.podcast
 
 import android.content.Context
 import android.os.Build
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.models.db.AppDatabase
 import au.com.shiftyjelly.pocketcasts.models.entity.Episode
 import au.com.shiftyjelly.pocketcasts.models.entity.Playlist
@@ -34,9 +36,18 @@ private const val CREATED_DEFAULT_PLAYLISTS = "createdDefaultPlaylists"
 class PlaylistManagerImpl @Inject constructor(
     private val settings: Settings,
     private val downloadManager: DownloadManager,
+    private val analyticsTracker: AnalyticsTrackerWrapper,
     @ApplicationContext private val context: Context,
     appDatabase: AppDatabase
 ) : PlaylistManager, CoroutineScope {
+
+    companion object {
+        const val ENABLED_KEY = "enabled"
+        const val LIMIT_KEY = "limit"
+        const val GROUP_KEY = "group"
+        const val SOURCE_KEY = "source"
+        const val SORT_ORDER_KEY = "sort_order"
+    }
 
     private val playlistDao = appDatabase.playlistDao()
 
@@ -139,7 +150,7 @@ class PlaylistManagerImpl @Inject constructor(
     override fun findEpisodes(playlist: Playlist, episodeManager: EpisodeManager, playbackManager: PlaybackManager): List<Episode> {
         val where = buildPlaylistWhere(playlist, playbackManager)
         val orderBy = getPlaylistOrderByString(playlist)
-        val limit = if (playlist.sortId == Playlist.PLAYLIST_SORT_LAST_DOWNLOAD_ATTEMPT_DATE) 1000 else 500
+        val limit = if (playlist.sortOrder() == Playlist.SortOrder.LAST_DOWNLOAD_ATTEMPT_DATE) 1000 else 500
         return episodeManager.findEpisodesWhere("$where ORDER BY $orderBy LIMIT $limit")
     }
 
@@ -150,7 +161,7 @@ class PlaylistManagerImpl @Inject constructor(
     }
 
     override fun observeEpisodes(playlist: Playlist, episodeManager: EpisodeManager, playbackManager: PlaybackManager): Flowable<List<Episode>> {
-        val limitCount = if (playlist.sortId == Playlist.PLAYLIST_SORT_LAST_DOWNLOAD_ATTEMPT_DATE) 1000 else 500
+        val limitCount = if (playlist.sortOrder() == Playlist.SortOrder.LAST_DOWNLOAD_ATTEMPT_DATE) 1000 else 500
         val queryAfterWhere = getPlaylistQuery(playlist, limit = limitCount, playbackManager = playbackManager)
         return episodeManager.observeEpisodesWhere(queryAfterWhere)
     }
@@ -160,16 +171,28 @@ class PlaylistManagerImpl @Inject constructor(
         return episodeManager.observeEpisodesWhere(queryAfterWhere)
     }
 
-    private fun getPlaylistOrderByString(playlist: Playlist): String? {
-        if (playlist.sortId == Playlist.PLAYLIST_SORT_NEWEST_TO_OLDEST || playlist.sortId == Playlist.PLAYLIST_SORT_OLDEST_TO_NEWEST) {
-            return "published_date " + (if (playlist.sortId == Playlist.PLAYLIST_SORT_NEWEST_TO_OLDEST) "DESC" else "ASC") +
-                ", added_date " + if (playlist.sortId == Playlist.PLAYLIST_SORT_NEWEST_TO_OLDEST) "DESC" else "ASC"
-        } else if (playlist.sortId == Playlist.PLAYLIST_SORT_SHORTEST_TO_LONGEST || playlist.sortId == Playlist.PLAYLIST_SORT_LONGEST_TO_SHORTEST) {
-            return "duration " + (if (playlist.sortId == Playlist.PLAYLIST_SORT_SHORTEST_TO_LONGEST) "ASC" else "DESC") + ", added_date DESC"
-        } else if (playlist.sortId == Playlist.PLAYLIST_SORT_LAST_DOWNLOAD_ATTEMPT_DATE) {
-            return "last_download_attempt_date DESC, published_date DESC"
+    private fun getPlaylistOrderByString(playlist: Playlist): String? = when (playlist.sortOrder()) {
+
+        Playlist.SortOrder.NEWEST_TO_OLDEST,
+        Playlist.SortOrder.OLDEST_TO_NEWEST -> {
+            "published_date " +
+                (if (playlist.sortOrder() == Playlist.SortOrder.NEWEST_TO_OLDEST) "DESC" else "ASC") +
+                ", added_date " +
+                if (playlist.sortOrder() == Playlist.SortOrder.NEWEST_TO_OLDEST) "DESC" else "ASC"
         }
-        return null
+
+        Playlist.SortOrder.SHORTEST_TO_LONGEST,
+        Playlist.SortOrder.LONGEST_TO_SHORTEST -> {
+            "duration " +
+                (if (playlist.sortOrder() == Playlist.SortOrder.SHORTEST_TO_LONGEST) "ASC" else "DESC") +
+                ", added_date DESC"
+        }
+
+        Playlist.SortOrder.LAST_DOWNLOAD_ATTEMPT_DATE -> {
+            "last_download_attempt_date DESC, published_date DESC"
+        }
+
+        else -> null
     }
 
     override fun create(playlist: Playlist): Long {
@@ -180,8 +203,12 @@ class PlaylistManagerImpl @Inject constructor(
         return id
     }
 
-    override fun update(playlist: Playlist) {
+    /**
+     * A null userPlayListUpdate parameter indicates that  the user did not initiate this update
+     */
+    override fun update(playlist: Playlist, userPlaylistUpdate: UserPlaylistUpdate?) {
         playlistDao.update(playlist)
+        sendPlaylistUpdateAnalytics(playlist, userPlaylistUpdate)
     }
 
     override fun updateAll(playlists: List<Playlist>) {
@@ -225,7 +252,10 @@ class PlaylistManagerImpl @Inject constructor(
         if (loggedIn) {
             playlist.deleted = true
             markAsNotSynced(playlist)
-            update(playlist)
+
+            // user initiated filter deletion, not update of any playlist properties, so this
+            // is not a user initiated update
+            update(playlist, userPlaylistUpdate = null)
         }
 
         if (!loggedIn) {
@@ -307,6 +337,57 @@ class PlaylistManagerImpl @Inject constructor(
 
     override fun findPlaylistsToSync(): List<Playlist> {
         return playlistDao.findNotSynced()
+    }
+
+    private fun sendPlaylistUpdateAnalytics(
+        playlist: Playlist,
+        userPlaylistUpdate: UserPlaylistUpdate?
+    ) {
+        // Don't send a filter updated event if the playlist is being created or if
+        // the user did not initiate the update
+        if (!playlist.draft && userPlaylistUpdate != null) {
+            userPlaylistUpdate.properties.map { playlistProperty ->
+                when (playlistProperty) {
+
+                    is FilterUpdatedEvent -> {
+                        val properties = mapOf(
+                            GROUP_KEY to playlistProperty.groupValue,
+                            SOURCE_KEY to userPlaylistUpdate.source.analyticsValue
+                        )
+                        analyticsTracker.track(AnalyticsEvent.FILTER_UPDATED, properties)
+                    }
+
+                    is PlaylistProperty.AutoDownload -> {
+                        val properties = mapOf(
+                            SOURCE_KEY to userPlaylistUpdate.source.analyticsValue,
+                            ENABLED_KEY to playlistProperty.enabled
+                        )
+                        analyticsTracker.track(AnalyticsEvent.FILTER_AUTO_DOWNLOAD_UPDATED, properties)
+                    }
+
+                    is PlaylistProperty.AutoDownloadLimit -> {
+                        val properties = mapOf(LIMIT_KEY to playlistProperty.limit)
+                        analyticsTracker.track(AnalyticsEvent.FILTER_AUTO_DOWNLOAD_LIMIT_UPDATED, properties)
+                    }
+
+                    is PlaylistProperty.Sort -> {
+                        val sortOrderString = when (playlistProperty.sortOrder) {
+                            Playlist.SortOrder.NEWEST_TO_OLDEST -> "newest_to_oldest"
+                            Playlist.SortOrder.OLDEST_TO_NEWEST -> "oldest_to_newest"
+                            Playlist.SortOrder.SHORTEST_TO_LONGEST -> "shortest_to_longest"
+                            Playlist.SortOrder.LONGEST_TO_SHORTEST -> "longest_to_shortest"
+                            Playlist.SortOrder.LAST_DOWNLOAD_ATTEMPT_DATE -> "last_download_attempt_date"
+                        }
+                        val properties = mapOf(SORT_ORDER_KEY to sortOrderString)
+                        analyticsTracker.track(AnalyticsEvent.FILTER_SORT_BY_CHANGED, properties)
+                    }
+
+                    PlaylistProperty.Color,
+                    PlaylistProperty.FilterName,
+                    PlaylistProperty.Icon -> { /* Do nothing. These are handled by the filter_edit_dismissed event. */ }
+                }
+            }
+        }
     }
 
     /**
@@ -513,7 +594,7 @@ class PlaylistManagerImpl @Inject constructor(
             audioVideo = Playlist.AUDIO_VIDEO_FILTER_ALL,
             allPodcasts = true,
             autoDownload = false,
-            sortId = Playlist.PLAYLIST_SORT_LAST_DOWNLOAD_ATTEMPT_DATE
+            sortId = Playlist.SortOrder.LAST_DOWNLOAD_ATTEMPT_DATE.value
         )
     }