+++ modules/services/views/src/main/java/au/com/shiftyjelly/pocketcasts/views/dialog/OptionsDialog.kt
@@ -119,4 +119,9 @@ class OptionsDialog : BottomSheetDialogFragment() {
             }
         }
     }
+
+    fun setOnDismiss(onDismiss: (() -> Unit)?): OptionsDialog {
+        this.onDismiss = onDismiss
+        return this
+    }
 }